# Improve `sexp_buffer_from_stream` performance