--- a/.gitignore
+++ b/.gitignore
@@ -3,3 +3,4 @@ __pycache__
 pip-wheel-metadata
 venv
 .venv
+.idea/